Fix: addStudentsProfilesByExcel
Some incorrect logic was corrected by restructuring the method of adding students by excel using as a base the already created method of adding students to a seedbed. Also, the controller method was restructured so that it only receives as parameter the file and the seedbed id, now it only calls the service method and does not perform logic.